pyproject.toml
[project]
name = "graph-weave"
version = "0.1.0"
description = "Use case for Graph and Weave for RAG"
authors = [
    {name = "Guandaline",email = "vhguandaline@gmail.com"}
]
license = {text = "Apache-2.0 license"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "opentelemetry-sdk (>=1.36.0,<2.0.0)",
    "opentelemetry-api (>=1.36.0,<2.0.0)",
    "opentelemetry-instrumentation-logging (>=0.57b0,<0.58)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "neo4j (>=5.28.2,<6.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "dynaconf (>=3.2.11,<4.0.0)",
    "weaviate (>=0.1.2,<0.2.0)",
    "weaviate-client (>=4.16.9,<5.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "opentelemetry-exporter-otlp-proto-grpc (>=1.36.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.57b0,<0.58)",
    "opentelemetry-instrumentation-requests (>=0.57b0,<0.58)",
    "opentelemetry-instrumentation-redis (>=0.57b0,<0.58)",
    "uvloop (>=0.21.0,<0.22.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "ruff (>=0.12.11,<0.13.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

